Fluent Validation Module - Step 2 - Extension Methods
