+++ samples/EndPoints/Base.Samples.EndPoints.WebApi/Program.cs
@@ -1,9 +1,9 @@
 using Base.Extensions.BackgroundWorker.Abstractions;
 using Base.Extensions.BackgroundWorker.KafkaConsumer;
+using Base.Infra.Validator;
 using Base.Sample.Application.KafkaServices;
 using Base.Sample.Application.People.Validators;
 using Base.Sample.BackgroundWorker.LocationService;
-using Base.Samples.EndPoints.WebApi.Validator;
 
 var builder = WebApplication.CreateBuilder(args);
 builder.Services.AddBaseApiCore("Base");
@@ -23,9 +23,8 @@ builder.Services.AddDbContext<BaseDbContext, SampleDbContext>(
         options.MigrationsAssembly(typeof(SampleDbContext).Assembly.GetName().Name);
     }));
 
-builder.Services.InitializeValidators(typeof(PersonInsertValidator));
 
-builder.Services.RegisterValidator<PersonInsertViewModel, PersonInsertValidator>();
+builder.Services.RegisterValidatorsByAssembly(typeof(PersonInsertViewModelValidator).Assembly, typeof(PersonInsertViewModel)?.Namespace ?? "");
 
 
 builder.Services.AddSwagger(builder.Configuration, "Swagger");
